requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
chromadb==0.4.18
PyPDF2==3.0.1
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
python-dotenv==1.0.0
huggingface-hub==0.19.4
torch==2.1.0
transformers==4.35.2
